calculates geocomplexity for vector data based on spatial dependence

This function calculates geocomplexity, a geospatial local complexity indicator, for variable in vector data based on spatial dependence.

Usage

geocd_vector(sfj, wt = NULL, normalize = TRUE)

Arguments

sfj

Vector object that can be converted to sf by sf::st_as_sf().

wt

(optional) Spatial weight matrix. Must be a matrix class. You can get a spatial weight matrix from spdep,rgeoda or tidyrgeoda package. If wt is not provided, geocomplexity will use a first-order queen adjacency binary matrix.

normalize

(optional) Whether to further normalizes the spatial local complexity. Default is TRUE.

Value

An sf object

Examples

library(sf)
#> Linking to GEOS 3.10.2, GDAL 3.4.1, PROJ 8.2.1; sf_use_s2() is TRUE
data("income")
inc = dplyr::select(income,Income)
gc = geocd_vector(inc)
gc
#> Simple feature collection with 333 features and 1 field
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: 112.9211 ymin: -43.63305 xmax: 153.6299 ymax: -9.223927
#> Geodetic CRS:  GDA94
#> # A tibble: 333 × 2
#>    Geocomplexiy_Income                                                  geometry
#>                  <dbl>                                        <MULTIPOLYGON [°]>
#>  1               0.663 (((149.979 -35.50246, 149.964 -35.50079, 149.9494 -35.55…
#>  2               0.642 (((148.8041 -35.71402, 148.8279 -35.72615, 148.8347 -35.…
#>  3               0.636 (((150.3754 -35.56524, 150.3647 -35.57078, 150.3226 -35.…
#>  4               0.671 (((149.0114 -33.93276, 149.0251 -33.93214, 149.0335 -33.…
#>  5               0.638 (((147.7137 -34.16162, 147.7441 -34.14091, 147.7769 -34.…
#>  6               0.668 (((151.485 -33.39868, 151.4689 -33.40979, 151.4439 -33.4…
#>  7               0.676 (((151.485 -33.39868, 151.4623 -33.39893, 151.4539 -33.3…
#>  8               0.656 (((149.323 -33.05916, 149.3436 -33.06254, 149.3543 -33.0…
#>  9               0.619 (((149.1264 -33.86642, 149.1198 -33.88783, 149.1048 -33.…
#> 10               0.673 (((150.5587 -32.75774, 150.5493 -32.77053, 150.557 -32.7…
#> # ℹ 323 more rows